move relevant tests to test_mosaic.py
